fix(ci): repair broken imports + stale test from the refactor merge
Three test-suite failures inherited from the merge-chessclient-refactor
merge (none in the student-vs-student feature itself):

- Pages/NewStudentProfile and Pages/piece-lessons/lesson-overlay imported
  assets via ../images/... instead of ../assets/images/... — corrected the
  paths (targets exist under src/assets/images/).
- PlayComputer.test asserted the old io('url', {...}) signature, but the
  refactor changed the source to io({ path, ... }) for the socket.io fix.
  Updated the expectation to match the corrected source.

react-scripts test now: 29 suites / 160 tests passing.
